xSize ySize floats

+++ Code/ComplexPlane.cpp
@@ -20,8 +20,8 @@ void ComplexPlane::updateRender()
 void ComplexPlane::zoomIn()
 {
     m_ZoomCount++;
-    double xSize = BASE_WIDTH * pow(BASE_ZOOM, m_ZoomCount);
-    double ySize = BASE_HEIGHT * m_aspectRatio * pow(BASE_ZOOM, m_ZoomCount);
+    float xSize = BASE_WIDTH * pow(BASE_ZOOM, m_ZoomCount);
+    float ySize = BASE_HEIGHT * m_aspectRatio * pow(BASE_ZOOM, m_ZoomCount);
     //m_plane_size(xSize, ySize); This doesn't work
     m_plane_size = sf::Vector2f(static_cast<float>(xSize), static_cast<float>(ySize));
     m_state = State::CALCULATING;
